# docs/_archive/2026-09-18/openapi.yaml
openapi: 3.1.0
info:
  title: YYC³ NovaMind API
  version: 2.0.0
  description: |
    YYC³ NovaMind 全栈 API 契约（OpenAPI 3.1）。

    本契约由 M3 MiniMax 导师于 2026-09-18 自动扫描 `app/api/**` 导出函数生成，
    覆盖 13 个路由文件 / 30 个端点，作为前后端对接的唯一权威来源。

    ## 五维契约原则
    - **时间维**：所有时间字段统一 ISO-8601 UTC
    - **空间维**：所有路径相对 baseURL（默认 `http://localhost:3218`）
    - **属性维**：所有响应含 `success: boolean`，错误码与 `lib/error-handler.ts` 对齐
    - **事件维**：所有 POST/PUT 操作记录 trace_id 用于审计
    - **关联维**：复杂资源附带 `_links` 字段指向关联资源
  contact:
    name: YYC³ Team
    email: admin@0379.email
  license:
    name: Copyright © 2025-2026 YanYuCloudCube
    identifier: YYC-Internal
servers:
  - url: http://localhost:3218
    description: 本地开发（端口 3218）
  - url: https://staging.api.yyc3.app
    description: 预发布
  - url: https://api.yyc3.app
    description: 生产
...
paths:
...
  /api/auth:
    get:
      tags: [Auth]
      summary: 查询当前会话
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Auth]
      summary: 登录 / 注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Auth]
      summary: 更新用户资料 / 刷新会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthUpdate'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ===== Chat =====
...
  /api/chat/stream:
    post:
      tags: [Chat]
      summary: AI 对话（流式 / SSE）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: SSE 流
          content:
            text/event-stream:
              schema:
                type: string
    get:
      tags: [Chat]
      summary: 获取流式连接元数据
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ===== Database =====
...
components:
  responses:
...
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          description: 业务是否成功
        data:
          description: 业务负载（任意 JSON）
        message:
          type: string
        trace_id:
          type: string
          description: 链路追踪 ID
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION, AUTH, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMIT, INTERNAL]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            severity:
              type: string
              enum: [LOW, MEDIUM, HIGH, CRITICAL]
            trace_id:
              type: string
...
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        action:
          type: string
          enum: [login, register, forgot_password]
    AuthUpdate:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        preferences:
          type: object
          additionalProperties: true
    ChatRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
        model:
          type: string
        temperature:
          type: number
        stream:
          type: boolean
          default: false
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: yyc3_session
      description: 会话 Cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 备用 Bearer Token
security:
  - sessionCookie: []
  - bearerAuth: []